Treat Go's int and uint as C's long and ulong
Otherwise there's no way to write cross bit-size Go code, since
C's int is always 32 bits while C's long equals the arch's bits.

+++ examples_test.go
@@ -12,12 +12,12 @@ func ExampleOpen_snprintf() {
 	if err != nil {
 		panic(err)
 	}
-	var snprintf func([]byte, uintptr, string, ...interface{}) int
+	var snprintf func([]byte, uint, string, ...interface{}) int
 	if err := dl.Sym("snprintf", &snprintf); err != nil {
 		panic(err)
 	}
 	buf := make([]byte, 200)
-	snprintf(buf, uintptr(len(buf)), "hello %s!\n", "world")
+	snprintf(buf, uint(len(buf)), "hello %s!\n", "world")
 	s := string(buf[:bytes.IndexByte(buf, 0)])
 	fmt.Println(s)
 	// Output: hello world!
